requirements.txt
uvicorn
fastapi
Pillow
opencv-python
torch
torchvision